requirements.txt
Django==5.0.6
djangorestframework==3.15.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
gunicorn==20.1.0
pytest
pytest-django
django-extensions